Added Vagrantfile and dev ansible playbook
Whew, that was quite a lot. Where do I start?

To utilize Vagrant, you first need to install virtualbox and vagrant
itself. After those are installed, simply do "vagrant up" from the
project's root directory. That's it! It will install and set everything
up for you. Once the playbook has completed provisioning, simply go to
localhost:8000 in your browser. This directory is shared with the
virtual machine, so all changes made here will be reflected live!

Give it a shot!