LICENSE: Add LibreJS Meta to included JavaScript
This patch adds all of the needed information to have all of the
JavaScript in the project be easily identifiable as free software by
both computer tools as well as humans. This is done by the following:

- Original sources are added and linked for all minified libraries.

- A data tag is added to the "About" link in the navbar menu which
alerts tools such as LibreJS to look on that page for the table of
includes mapped to their respective licenses.

- A table is added to the "About" page which includes links to each
included file, links to their respective licenses, and then a link to a
decompressed version of each.

Signed-off-by: Mark Stenglein <mark@stengle.in>