LibreJS: Fix Table Formatting
This commit makes the LibreJS table utilize the bootstrap styles
make the table not look absurdly ugly. It also adds table headers,
which is nice :ok_hand:.

In the future, it might be neat to add a test case which loads the
LibreJS parse engine to check for licenses that are not added to the
table. Is there a way to load the IceCat addin in some kind of
headless firefox?

Signed-off-by: Mark Stenglein <mark@stengle.in>